message

+++ Suppression.php
@@ -14,11 +14,11 @@ error_reporting(E_ALL);
     if(isset($_COOKIE['id']) && isset($_COOKIE['token']) && isset($_POST['supprimer'])){
         $sql = "DELETE from Article where idArticle= '$idArticle'";
         $img = "SELECT url from Article where idArticle= '$idArticle'";
-        //mysql_query($img)) or die('Erreur SQL !'.$sql.'<br>'.mysql_error());
+        mysql_query($img)) or die('Erreur SQL !'.$img.'<br>'.mysql_error());
         mysql_query($sql)) or die('Erreur SQL !'.$sql.'<br>'.mysql_error());
-        /*if($sql==1){
+        if($sql==1){
             unlink(realpath($img));
-        };*/
+        };
     }
     mysql_close();  // on ferme la connexion 
 ?>
